+++ CHANGES.rst
@@ -7,6 +7,10 @@ Changelog of dask-geomodeling
 
 - Added GeometryWKTSource.
 
+- Reworked the docstrings of all rasterblocks.
+
+- Renamed the 'location' parameter of raster.misc.Step to 'value'.
+
 
 2.2.1 (2020-02-04)
 ------------------